Better money earning
1) made on jobs
2) can be fractional per second

+++ Assets/Scripts/Data/Player/IncomeSource.cs
@@ -4,6 +4,6 @@ namespace Data.Player
 {
 	public struct IncomeSource : IComponentData
 	{
-		public int GoldPerSecond;
+		public float GoldPerSecond;
 	}
 }
